Simpler approach to the fix using constraints
