+++ revenuecatui/src/iosMain/kotlin/com/revenuecat/purchases/kmp/ui/revenuecatui/UIKitCustomerCenter.kt
@@ -24,10 +24,10 @@ internal fun UIKitCustomerCenter(
     // Intrinsic content size from UIKit
     var intrinsicContentSizePx by remember { mutableStateOf(0) }
 
-    // We remember this wrapper so we can keep a reference to ConstrainingViewController, even
+    // We remember this wrapper so we can keep a reference to CustomerCenterUIViewController, even
     // during recompositions. CustomerCenterUIViewController itself is not yet instantiated here.
     val viewControllerWrapper = remember {
-        ViewControllerWrapper<ConstrainingViewController<CustomerCenterUIViewController>>(null)
+        ViewControllerWrapper<CustomerCenterUIViewController>(null)
     }
 
     // Keep a reference to IosCustomerCenterDelegate across recompositions
@@ -41,7 +41,7 @@ internal fun UIKitCustomerCenter(
                 constraints.copy(minHeight = intrinsicContentSizePx)
             else constraints
 
-            viewControllerWrapper.wrapped?.constraints = constraintsToUse
+            viewControllerWrapper.applyConstraints(constraintsToUse, density)
             val placeable = measurable.measure(constraintsToUse)
 
             layout(placeable.width, placeable.height) {
@@ -49,16 +49,18 @@ internal fun UIKitCustomerCenter(
             }
         },
         factory = {
-            val customerCenterViewController = CustomerCenterUIViewController()
+            CustomerCenterUIViewController()
                 .apply {
-                setDelegate(delegate)
-                setOnCloseHandler(onDismiss)
-                view.getIntrinsicContentSizeOfFirstSubView()
-                    ?.also { intrinsicContentSizePx = with(density) { it.dp.roundToPx() } }
-            }
-
-            ConstrainingViewController(customerCenterViewController, density)
-                .also { viewControllerWrapper.wrapped = it }
+                    setDelegate(delegate)
+                    setOnCloseHandler(onDismiss)
+                    view.getIntrinsicContentSizeOfFirstSubView()
+                        ?.also {
+                            intrinsicContentSizePx = with(density) { it.dp.roundToPx() }
+                            println("TESTING [paywallViewController] intrinsicContentSizePx: $intrinsicContentSizePx")
+                        }
+                }.also {
+                    viewControllerWrapper.wrapped = it
+                }
         },
         properties = uiKitInteropPropertiesNonExperimental(
             nonCooperativeInteractionMode = true,
